requirements.txt
# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0
dirtyjson>=1.0.8
beautifulsoup4>=4.12.0

# Cloud & AWS
boto3>=1.28.0
awswrangler>=3.0.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
PyYAML>=6.0
seleniumbase>=4.48.0
dukascopy-python>=0.2.0